feat: allow version pinning if install_mode package_manager
This could be considered a breaking change since it moves the package_manager install mode to use the apt module instead of generic package module, but mergerfs isn't available in most, if not all, RHEL-based first-party package repositories.
Perhaps in the future this could be expanded to add a trusted third-party package repository if the user requests this install mode when using a RHEL-based distro. For now, it will just warn and exit.