fix: remove localhost delegation for GitHub query
Fixes #26.

The delegation to localhost was seen to cause the role to hang at this step for reasons unknown.

Since we aren't directly querying the API anyway, this wasn't of great value. If we had been, it would've been a nice way of reducing the number of API hits when targeting multiple hosts with the role, but I digress.